+++ app/Models/Anggaran.php
@@ -16,13 +16,14 @@ class Anggaran extends Model
         'keterangan_lainnya', 
         'nilai_anggaran', 
         'verifikasi', 
-        'status'
+        'status',
+        'nilai_realisasi',
     ];
 
     public function detail_norekening()
-    {
-        return $this->belongsTo('App\Models\DetailNorekening'); // Ensure class name matches convention
-    }
+{
+    return $this->belongsTo('App\Models\Detail_Norekening', 'detail_norekening_id');
+}
 
     // Function to toggle verification status
     public function toggleVerifikasi()